Add read-repairs test and misc bug fixes
This patch adds to the testing infra to make it possible to simulate
partial writes and uses it to test read repairs.

Other misc changes:
 - Move struct definition out of .h file (Bad practice to have it in.h)
 - Populate (struct node*)->is_local field for responses from the data
   store since it's not pre-populated. (TODO: Move logic to diff place)
