+++ test/dual_run.py
@@ -17,9 +17,13 @@ class ResultMismatchError(Exception):
         return ret
 
 class dual_run():
-    def __init__(self, r, d, debug=None):
-        self.r = r
-        self.d = d
+    def __init__(self, standalone_redis, dyno_cluster, debug=None):
+        self.standalone_redis = standalone_redis
+        self.dyno_cluster = dyno_cluster
+
+        self.redis_conn = standalone_redis.get_connection()
+        self.dyno_conn = dyno_cluster.get_connection()
+
         self.debug = debug
         self.sort_before_cmp = False
 
@@ -29,11 +33,19 @@ class dual_run():
     def set_sort_before_compare(self, should_sort):
         self.sort_before_cmp = should_sort
 
+    # Returns the underlying DynoCluster object
+    def get_dynomite_cluster(self):
+        return self.dyno_cluster
+
+    def ensure_underlying_dyno_conn_is_multi_dc(self):
+        self.dyno_conn = self.dyno_cluster.get_connection_to_multi_rack_dc()
+        assert self.dyno_conn != None , "Could not obtain connection to multi-rack DC"
+
     def run_verify(self, func, *args):
         r_result = None
         d_result = None
-        r_func = getattr(self.r, func)
-        d_func = getattr(self.d, func)
+        r_func = getattr(self.redis_conn, func)
+        d_func = getattr(self.dyno_conn, func)
         r_result = r_func(*args)
         i = 0
         retry_limit = 3
@@ -65,7 +77,7 @@ class dual_run():
 
     def run_dynomite_only(self, func, *args):
         d_result = None
-        d_func = getattr(self.d, func)
+        d_func = getattr(self.dyno_conn, func)
         i = 0
         retry_limit = 3
         while i < retry_limit:
@@ -85,3 +97,12 @@ class dual_run():
             print("Query: %s %s" % (func, str(args)))
             print("Dyno result: %s" % str(d_result))
         return d_result
+
+    def run_redis_only(self, func, *args):
+        r_result = None
+        r_func = getattr(self.redis_conn, func)
+        r_result = r_func(*args)
+        if self.debug:
+            print("Query: %s %s" % (func, str(args)))
+            print("Redis result: %s" % str(r_result))
+        return r_result
